requirements.txt
Flask==3.1.2
pandas==2.3.3
psycopg2==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.30
Flask-SQLAlchemy==3.1.1
